convert to long

+++ main.cu
@@ -24,7 +24,7 @@ using std::cerr;
 using std:: endl;
 
 __global__
-void cudaMultiply(int *array1, int *array2, int *array3, int size) {
+void cudaMultiply(long *array1, long *array2, long *array3, long size) {
     int index = blockIdx.x * blockDim.x + threadIdx.x;
     int stride = blockDim.x * gridDim.x;
     for (int i = index; i < size; i += stride)
@@ -38,15 +38,15 @@ int main(int argc, char **argv) {
     return 1;
     #endif
 
-    long int arraySize = ARRAYSIZE;
+    long arraySize = ARRAYSIZE;
     cout << "ArraySize: " << arraySize << endl;
     cout << "NumTries: " << NUMTRIES << endl;
     int threadArray[] = {1, 2, 4, 6, 8, 12, 16};
     int threadArraySize = sizeof(threadArray)/sizeof(threadArray[0]);
 
-    long int* hostArray1 = new int[arraySize];
-    long int* hostArray2 = new int[arraySize];
-    long int* hostArray3 = new int[arraySize];
+    long* hostArray1 = new long int[arraySize];
+    long* hostArray2 = new long int[arraySize];
+    long* hostArray3 = new long int[arraySize];
 
     for(int i = 0; i < arraySize; i++) {
         hostArray1[i] = rand() % 100;
@@ -71,7 +71,7 @@ int main(int argc, char **argv) {
         cout << "Time to complete operation with " << threadArray[i] << " threads(ms): " << averageTime * 1000 << endl;
     }
 
-    long int *deviceArray1, *deviceArray2, *deviceArray3;
+    long *deviceArray1, *deviceArray2, *deviceArray3;
     cudaMallocManaged(&deviceArray1, arraySize * sizeof(int));
     cudaMallocManaged(&deviceArray2, arraySize * sizeof(int));
     cudaMallocManaged(&deviceArray3, arraySize * sizeof(int));
